# photodome-ios/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/v1/events": {
      "post": {
        "operationId": "createEvent",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostEventAccessDto"
                }
              }
            }
          }
        },
        "summary": "Create an event without an account",
        "tags": [
          "events"
        ]
      }
    },
    "/v1/events/join": {
      "post": {
        "operationId": "joinEvent",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventAccessDto"
                }
              }
            }
          },
          "404": {
            "description": "Invite invalid, rotated, expired, or unavailable."
          }
        },
        "summary": "Join a private event with its short code",
        "tags": [
          "events"
        ]
      }
    },
...
    "/v1/events/{eventId}/photos": {
      "get": {
        "operationId": "listEventPhotos",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumPhotoPageDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "eventCapability": []
          }
        ],
        "summary": "List ready photos with short-lived private read URLs",
        "tags": [
          "photos"
        ]
      }
    },
    "/v1/events/{eventId}/photos/reservations": {
      "post": {
        "operationId": "reservePhotoUpload",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePhotoReservationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoUploadGrantDto"
                }
              }
            }
          },
          "409": {
            "description": "Uploads are closed or the event photo limit was reached."
          }
        },
        "security": [
          {
            "eventCapability": []
          }
        ],
        "summary": "Reserve a photo and create a direct resumable upload session",
        "tags": [
          "photos"
        ]
      }
    },
    "/v1/events/{eventId}/photos/{photoId}/upload-session": {
      "post": {
        "operationId": "renewPhotoUploadSession",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "photoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoUploadGrantDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "eventCapability": []
          }
        ],
        "summary": "Reissue a resumable session for an owned reservation",
        "tags": [
          "photos"
        ]
      }
    },
    "/v1/events/{eventId}/photos/{photoId}/complete": {
      "post": {
        "operationId": "completePhotoUpload",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "photoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoCompletionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "eventCapability": []
          }
        ],
        "summary": "Verify an uploaded object and enqueue private image processing",
        "tags": [
          "photos"
        ]
      }
    },
    "/v1/events/{eventId}/photos/{photoId}": {
      "delete": {
        "operationId": "removeEventPhoto",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "photoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "eventCapability": []
          }
        ],
        "summary": "Delete a ready photo contributed by the current member",
        "tags": [
          "photos"
        ]
      }
    },
...
  },
  "info": {
    "title": "PhotoDome API",
    "description": "Backend API for the PhotoDome iPhone app.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "eventCapability": {
        "scheme": "bearer",
        "bearerFormat": "event capability",
        "type": "http",
        "description": "Opaque event-scoped host or guest capability."
      }
    },
    "schemas": {
      "CreateEventDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "displayName": {
            "type": "string",
            "example": "James",
            "minLength": 1,
            "maxLength": 50
          },
          "locationLabel": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          }
        },
        "required": [
          "name",
          "displayName"
        ]
      },
      "EventViewerDto": {
        "type": "object",
        "properties": {
          "memberId": {
            "type": "string",
            "format": "uuid"
          },
          "role": {
            "type": "string",
            "enum": [
              "HOST",
              "GUEST"
            ]
          }
        },
        "required": [
          "memberId",
          "role"
        ]
      },
      "EventSnapshotDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "James's birthday"
          },
          "hostDisplayName": {
            "type": "string",
            "example": "James"
          },
          "locationLabel": {
            "type": "string",
            "nullable": true,
            "example": "Shibuya"
          },
          "state": {
            "type": "string",
            "enum": [
              "LIVE",
              "ENDED",
              "EXPIRING"
            ]
          },
          "memberCount": {
            "type": "number",
            "minimum": 1,
            "maximum": 100,
            "example": 2
          },
          "readyPhotoCount": {
            "type": "number",
            "minimum": 0,
            "example": 24
          },
          "createdAt": {
            "type": "string",
            "example": "2026-07-25T00:00:00.000Z"
          },
          "endedAt": {
            "type": "string",
            "nullable": true
          },
          "expiresAt": {
            "type": "string",
            "nullable": true
          },
          "uploadsRestrictedAt": {
            "type": "string",
            "nullable": true
          },
          "viewer": {
            "$ref": "#/components/schemas/EventViewerDto"
          }
        },
        "required": [
          "id",
          "name",
          "hostDisplayName",
          "state",
          "memberCount",
          "readyPhotoCount",
          "createdAt",
          "viewer"
        ]
      },
      "HostEventAccessDto": {
        "type": "object",
        "properties": {
          "event": {
            "$ref": "#/components/schemas/EventSnapshotDto"
          },
          "capability": {
            "type": "string",
            "example": "pdc_OpaqueEventCapabilityReturnedOnce",
            "writeOnly": true
          },
          "joinCode": {
            "type": "string",
            "example": "7JMPK4QX"
          }
        },
        "required": [
          "event",
          "capability",
          "joinCode"
        ]
      },
      "JoinEventDto": {
        "type": "object",
        "properties": {
          "joinCode": {
            "type": "string",
            "example": "7JMP-K4QX",
            "minLength": 8,
            "maxLength": 16
          },
          "displayName": {
            "type": "string",
            "example": "Taylor",
            "minLength": 1,
            "maxLength": 50
          }
        },
        "required": [
          "joinCode",
          "displayName"
        ]
      },
      "EventAccessDto": {
        "type": "object",
        "properties": {
          "event": {
            "$ref": "#/components/schemas/EventSnapshotDto"
          },
          "capability": {
            "type": "string",
            "example": "pdc_OpaqueEventCapabilityReturnedOnce",
            "writeOnly": true
          }
        },
        "required": [
          "event",
          "capability"
        ]
      },
...
      "Object": {
        "type": "object",
        "properties": {}
      },
      "AlbumPhotoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "contributorMemberId": {
            "type": "string",
            "format": "uuid"
          },
          "width": {
            "type": "number"
          },
          "height": {
            "type": "number"
          },
          "capturedAt": {
            "type": "string",
            "nullable": true
          },
          "readyAt": {
            "type": "string"
          },
          "displayUrl": {
            "type": "string"
          },
          "thumbnailUrl": {
            "type": "string"
          },
          "urlsExpireAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "contributorMemberId",
          "width",
          "height",
          "capturedAt",
          "readyAt",
          "displayUrl",
          "thumbnailUrl",
          "urlsExpireAt"
        ]
      },
      "AlbumPhotoPageDto": {
        "type": "object",
        "properties": {
          "photos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AlbumPhotoDto"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          },
          "readyPhotoCount": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "photos",
          "nextCursor",
          "readyPhotoCount"
        ]
      },
      "CreatePhotoReservationDto": {
        "type": "object",
        "properties": {
          "contentType": {
            "type": "string",
            "enum": [
              "image/jpeg"
            ]
          },
          "byteSize": {
            "type": "number",
            "minimum": 1,
            "maximum": 20971520
          },
          "sha256": {
            "type": "string",
            "pattern": "^[a-fA-F0-9]{64}$",
            "description": "SHA-256 of the exact prepared upload bytes."
          },
          "width": {
            "type": "number",
            "minimum": 1,
            "maximum": 20000
          },
          "height": {
            "type": "number",
            "minimum": 1,
            "maximum": 20000
          },
          "capturedAt": {
            "type": "string"
          },
          "orientation": {
            "type": "number",
            "minimum": 1,
            "maximum": 8,
            "default": 1
          }
        },
        "required": [
          "contentType",
          "byteSize",
          "sha256",
          "width",
          "height"
        ]
      },
      "PhotoUploadGrantDto": {
        "type": "object",
        "properties": {
          "photoId": {
            "type": "string",
            "format": "uuid"
          },
          "uploadUrl": {
            "type": "string",
            "description": "Bearer-secret GCS resumable session URI. Never log or persist outside the upload queue."
          },
          "contentType": {
            "type": "string",
            "enum": [
              "image/jpeg"
            ]
          },
          "byteSize": {
            "type": "number"
          },
          "state": {
            "type": "string",
            "enum": [
              "RESERVED",
              "PROCESSING",
              "READY",
              "FAILED",
              "REMOVED",
              "EXPIRED"
            ]
          }
        },
        "required": [
          "photoId",
          "uploadUrl",
          "contentType",
          "byteSize",
          "state"
        ]
      },
      "PhotoCompletionDto": {
        "type": "object",
        "properties": {
          "state": {
            "type": "string",
            "enum": [
              "RESERVED",
              "PROCESSING",
              "READY",
              "FAILED",
              "REMOVED",
              "EXPIRED"
            ]
          }
        },
        "required": [
          "state"
        ]
      },
...
    }
  }
}